requirements.txt
streamlit
pandas
playwright
beautifulsoup4
python-dotenv
google-generativeai
fastapi
uvicorn
pydantic